CurrencyConvert waiting fix on datalog
